Pk registration works

+++ src/shared/Messages.h
@@ -1,7 +1,7 @@
 /** Represents a message from a Chit client to another Chit client */
 typedef struct {
     enum { MESSAGE } message_type; /* Message type */
-    unsigned int timestamp; /* Timestamp of the sent message */
+    unsigned long int timestamp; /* Timestamp of the sent message */
     unsigned int user_id; /* Sender ID of the message */
     unsigned long *payload; /* Encrypted message contents */
 } Chit_Message;
@@ -9,7 +9,7 @@ typedef struct {
 /** Represents a message to or from the Public Key Server */
 typedef struct {
     enum {FETCH_PK, FETCH_PK_ACK, REGISTER_PK, REGISTER_PK_ACK} message_type; /* Message type */
-    unsigned int timestamp; /* Timestamp of the sent message */
+    unsigned long int timestamp; /* Timestamp of the sent message */
     unsigned int user_id; /* Id of the user who sent the message, if applicable */
 
     /* If fetch_pk, this is requested user id for pk request */
@@ -25,7 +25,7 @@ typedef struct {
 /** Represents a message to or from the Address Server */
 typedef struct {
     enum {REGISTER_ADDR, REGISTER_ADDR_ACK, FETCH_ADDR, FETCH_ADDR_ACK} message_type; /* Message type */
-    unsigned int timestamp; /* Timestamp of the sent message */
+    unsigned long int timestamp; /* Timestamp of the sent message */
     unsigned int user_id; /* Id of the user who sent the message, if applicable */
     unsigned int user_port; /* Port address of the chit client in the case of REGISTER_ACK */
     char* user_addr; /* IP address of the chit client in the case of FETCH_ADDR */
